graph debug logging improvements
